--- app/src/main/java/com/example/easymathforpreschool/Numbers/Three.java
@@ -1,4 +1,0 @@
-package com.example.easymathforpreschool.Numbers;
-
-public class Three {
-}
