+++ app/src/main/java/com/example/easymathforpreschool/numbers/Six.java
@@ -1,0 +1,4 @@
+package com.example.easymathforpreschool.numbers;
+
+public class Six {
+}
