+++ app/src/main/java/com/example/easymathforpreschool/numbers/Three.java
@@ -1,0 +1,4 @@
+package com.example.easymathforpreschool.numbers;
+
+public class Three {
+}
