+++ app/src/main/java/com/example/easymathforpreschool/numbers/Two.java
@@ -1,0 +1,4 @@
+package com.example.easymathforpreschool.numbers;
+
+public class Two {
+}
